requirements.txt
terra_sdk
discord.py